Add database factories & seeders to all models. Create techSkills ENUM and refactor some migrations

+++ database/migrations/2024_07_20_065111_create_universities_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('universities', function (Blueprint $table) {
             $table->id();
             $table->string('name', 200)->unique();;
-            $table->integer('evaluation');
+            $table->float('evaluation');
             $table->timestamps();
         });
     }
